% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{rg_gdal_capabilities}
\alias{rg_gdal_capabilities}
\title{Query GDAL runtime driver capabilities}
\usage{
rg_gdal_capabilities(format)
}
\arguments{
\item{format}{Character scalar (e.g. `"GTiff"` or `"COG"`).}
}
\value{
A list with elements:
\itemize{
  \item version: GDAL version string
  \item driver: Driver short name
  \item has_create: TRUE/FALSE if `GDALCreate()` works
  \item has_createcopy: TRUE/FALSE if `GDALCreateCopy()` is supported
  \item has_virtualio: TRUE/FALSE if /vsimem/ supported
  \item datatypes: Character vector of supported GDAL types
}
}
\description{
Query GDAL runtime driver capabilities
}
\examples{
rg_gdal_capabilities("COG")
rg_gdal_capabilities("GTiff")
}
