% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{rg_info}
\alias{rg_info}
\title{Inspect Raster Metadata}
\usage{
rg_info(path)
}
\arguments{
\item{path}{Path to raster dataset.}
}
\value{
A list containing `width`, `height`, `bands`, `dtype`, `gt`, `crs`,
  `nodata`, and logical flags indicating presence of `color_table` and `categories`.
}
\description{
Retrieve basic metadata about a raster dataset using GDAL.
}
